refactor things into core
