+++ lib/core.js
@@ -1,95 +1,48 @@
 
-var es = require('event-stream')
-var EventEmitter = require('events').EventEmitter;
+var es = require('event-stream');
+var dispatcher = require('./dispatcher');
   ;
-// TODO: move much of this to ./dispatcher.js
 
-function sniff (data) {
-  var type = 'unknown';
-  if (data.type) {
-    type = data.type;
-  } else {
-    if ('sgv' in data) {
-      type = 'sgv';
-    }
-    if ('mbg' in data) {
-      type = 'mbg';
-    }
-  }
-  return type;
+function init (ctx) {
+  ctx.inputs.on('data', console.log.bind(console, 'DEBUG DATA'));
+  ctx.inputs.on('sgv', console.log.bind(console, 'DEBUG SGV DATA'));
+  ctx.inputs.on('/download/protobuf', console.log.bind(console, 'DEBUG DOWNLOAD DATA'));
+  ctx.inputs.write({msg: 'hello world'});
+
+  ctx.heartbeat.pipe(ctx.inputs);
+  ctx.heartbeat.uptime( );
 }
 
-function hash (data) {
-  var type = data.type;
-  var list = [ ];
-  var prefix = (type || 'datum') + '://';
-  var key = prefix + list.join('/');
-  return key;
+function core (env, ctx) {
+  var events = ctx.inputs = dispatcher(env, ctx);
+  var store = ctx.store;
+
+  ctx.pushover = require('./pushover')(env);
+  ctx.entries = require('./entries')(env.mongo_collection, ctx.store, ctx.inputs);
+  ctx.settings = require('./settings')(env.settings_collection, ctx.store, ctx.inputs);
+  ctx.treatments = require('./treatments')(env.treatments_collection, ctx.store, ctx.pushover, ctx.inputs);
+  ctx.devicestatus = require('./devicestatus')(env.devicestatus_collection, ctx.store, ctx.inputs);
+  ctx.pebble = require('./pebble');
+  console.info("Ensuring indexes");
+
+  store.ensureIndexes(ctx.entries( ), ctx.entries.indexedFields);
+  store.ensureIndexes(ctx.treatments( ), ctx.treatments.indexedFields);
+  store.ensureIndexes(ctx.devicestatus( ), ctx.devicestatus.indexedFields);
+
+
+  ctx.heartbeat = require('./ticker')(env, ctx)
+     ;
+     // .pipe(ctx.inputs);
+
+  if (env.MQTT_MONITOR) {
+    var mqtt = require('./mqtt')(env, ctx.inputs);
+    var es = require('event-stream');
+    // es.pipeline(mqtt.entries, ctx.entries.map( ), mqtt.every(ctx.entries));
+  }
+  events.init = init;
+  return events;
 }
 
-function hook (data, stream) {
-  var ev = es.through(function (item) {
-    this.push(item);
-  });
-  ev.name = data.type;
-  ev.value = data;
-  return ev;
-}
-
-function dispatcher (opts, fn) {
-  function default_cb (h, next) { next(null, h); }
-  var cb = (fn && fn.call) ? fn : (opts && opts.call ? opts : default_cb);
-  function writer (data) {
-    var type = sniff(data);
-    if (!data.type) {
-      data.type = type;
-    }
-    /*
-    TODO: enforce hash/sig somewhere, but not here, rather delegate to
-    any known models, then do last resort.
-    var sig = hash(data);
-    if (!data.sig) {
-      data.sig = sig;
-    }
-    */
-    var ev = hook(data, this);
-    this.push(ev);
-  }
-  function ender ( ) {
-    console.log('ending');
-  }
-  var stream = es.through(writer, ender);
-
-  function emitter ( ) { }
-
-  function start (ev, next) {
-    // TODO: inspect known models/register, bless with mixins?
-  }
-
-  function finish (ev, next) {
-    // TODO: execute hook's own stream, using one writable to process
-    // single data entry
-    // es.pipeline(es.readArray([hook.value]), hook, es.writeArray(done));
-    /*
-    function done (err, results) {
-      // next(err, results);
-      // stream.emit('finished', results, ...);
-      // next(err);
-    }
-    */
-  }
-  function iter (ev, next) {
-    // TODO:
-    // hook.inject(stream);
-    stream.emit(ev.name, ev);
-    next(null, ev);
-  }
-
-  es.pipeline(stream, es.map(iter));
-  return stream;
-}
-
-
-module.exports = dispatcher;
+module.exports = core;
